Refactor: 네이밍수정 , const 분리
