+++ src/constants.js
@@ -1,2 +1,6 @@
-export const LOTTORANGE=45;   //로또숫자범위
-export const LOTTOLENGTH= 6;
+export const LOTTORANGE = 45;   //로또 숫자범위
+export const LOTTOLENGTH = 6;  //로또 길이
+export const LOTTOPRICE = 1000; //로또가격
+
+
+export const WINMONEY=[5000,50000,1500000,30000000,2000000000]; //당첨금액
